+++ src/periodogram/sin_cos_iterator.rs
@@ -1,6 +1,71 @@
-//! Recurrent sine-cosine implementation
+//! Sine-cosine iterator implementations
 
 use crate::float_trait::Float;
+
+pub enum SinCosIterator<'a, T> {
+    LinearGrid(RecurrentSinCos<T>),
+    Values(ValueSinCosIterator<'a, T>),
+}
+
+impl<'a, T: Float> SinCosIterator<'a, T> {
+    pub fn from_angles(iter: impl Iterator<Item = T> + 'a) -> Self {
+        Self::Values(ValueSinCosIterator::new(iter))
+    }
+
+    pub fn from_linear_grid(first: T, step: T) -> Self {
+        let rec_sin_cos = RecurrentSinCos::new(first, step);
+        Self::LinearGrid(rec_sin_cos)
+    }
+}
+
+impl<'a, T> Iterator for SinCosIterator<'a, T>
+where
+    T: Float,
+{
+    type Item = (T, T);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        match self {
+            SinCosIterator::LinearGrid(iter) => iter.next(),
+            SinCosIterator::Values(iter) => iter.next(),
+        }
+    }
+}
+
+pub struct ValueSinCosIterator<'a, T> {
+    value_iter: Box<dyn Iterator<Item = T> + 'a>,
+}
+
+impl<'a, T> ValueSinCosIterator<'a, T> {
+    pub fn new(value_iter: impl Iterator<Item = T> + 'a) -> Self {
+        Self {
+            value_iter: Box::new(value_iter),
+        }
+    }
+}
+
+impl<'a, T> Iterator for ValueSinCosIterator<'a, T>
+where
+    T: Float,
+{
+    type Item = (T, T);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.value_iter.next().map(T::sin_cos)
+    }
+}
+
+impl<T> From<RecurrentSinCos<T>> for SinCosIterator<'static, T> {
+    fn from(sin_cos: RecurrentSinCos<T>) -> Self {
+        Self::LinearGrid(sin_cos)
+    }
+}
+
+impl<'a, T: Float> From<ValueSinCosIterator<'a, T>> for SinCosIterator<'a, T> {
+    fn from(sin_cos: ValueSinCosIterator<'a, T>) -> Self {
+        Self::Values(sin_cos)
+    }
+}
 
 /// Iterator over sin(start + k*step), cos(start + k*step) pairs
 ///