+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Breaking** Periodogram `FreqGrid` is refactored, now Periodogram uses new
   `FreqGridStrategy` https://github.com/light-curve/light-curve-feature/pull/205 https://github.com/light-curve/light-curve-feature/pull/206
+- `FreqGrid` now supports "fixed" grids, including an arbitrary sorted
+  array https://github.com/light-curve/light-curve-feature/pull/207
 
 ### Deprecated